+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Common
 
    * item states can be kept in the crash-free registry
    * controller PIDs in "test" command (requires master key)
+   * launch pptop directly from eva-shell
 
 3.4.0 (2021-07-07)
 ==================