+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Common
 ------
 
    * fix: recombined ACLs speed tuning
+   * fix: db updates on-exit correct transactions commit/rollback
 
    * item states can be kept in the crash-free registry
    * controller PIDs in "test" command (requires master key)